+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog][keepachangelog] and this project adher
 
 - Maximal `illuminate/*` packages version now is `7.*`
 - Minimal `PHP` version now is `^7.2`
+- Class `AvtoDev\VehicleLogotypes\VehicleLogotypes` finalized
 
 ### Added